feat/#33: 핀 개수 수정
