+++ src/components/board/PinGallery.jsx
@@ -1,13 +1,27 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import PinGalleryImg from './PinGalleryImg';
 import PinGalleryTitle from './PinGalleryTitle';
+import { getUserPins } from '../../services';
+import { useParams } from 'react-router-dom';
 
 function PinGallery() {
+  const { uid } = useParams();
+  const [pin, setPin] = useState(null);
+
+  const getPinData = async () => {
+    const { data } = await getUserPins(uid);
+    setPin(data.pin);
+  };
+  useEffect(() => {
+    getPinData();
+  }, []);
+
+  if (!pin) return;
   return (
     <StyledRoot>
-      <PinGalleryTitle />
-      <PinGalleryImg />
+      <PinGalleryTitle pin={pin} />
+      <PinGalleryImg pin={pin} />
     </StyledRoot>
   );
 }
